clients/python/pyproject.toml
[project]
name = "gvdb"
version = "0.7.0" # x-release-please-version
description = "Python client for GVDB distributed vector database"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.urls]
Homepage = "https://github.com/JonathanBerhe/gvdb"
Repository = "https://github.com/JonathanBerhe/gvdb"

[dependency-groups]
dev = [
    "grpcio-tools>=1.60.0",
    "pytest>=7.0",
    "ruff>=0.15.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gvdb"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
exclude = ["gvdb/pb/"]
